hegarty/sora_interface.py
"""
SoraInterface: Interface for Sora-2 video generation
"""

import logging
import time
import base64
from typing import Optional, Dict, Any
import numpy as np
from PIL import Image
import io
import json

logger = logging.getLogger(__name__)


class SoraInterface:
    """
    Interface for Sora-2 video generation API.
    
    This class handles communication with Sora-2 for generating mental rotation videos.
    If Sora API is not available, it provides a simulation mode for development.
    """
    
    def __init__(self, api_key: Optional[str] = None, config: Optional[Any] = None):
        """
        Initialize Sora interface.
        
        Args:
            api_key: Sora API key (optional, uses simulation if not provided)
            config: Configuration object
        """
        self.api_key = api_key
        self.config = config
        self.simulation_mode = not api_key
        
        if self.simulation_mode:
            logger.warning("Sora API key not provided. Running in simulation mode.")
        else:
            logger.info("Sora interface initialized with API access")
    
    def generate_video(
        self,
        prompt: str,
        image: str,
        duration: int = 3,
        fps: int = 10,
        resolution: str = "1024x1024"
    ) -> Dict[str, Any]:
        """
        Generate a video showing mental rotation or perspective change.
        
        Args:
            prompt: Text prompt describing the transformation
            image: Base64 encoded starting image
            duration: Video duration in seconds
            fps: Frames per second
            resolution: Output resolution
        
        Returns:
            Dictionary containing video data and metadata
        """
        logger.info(f"Generating video: {prompt[:50]}...")
        
        if self.simulation_mode:
            return self._simulate_video_generation(
                prompt, image, duration, fps, resolution
            )
        else:
            return self._call_sora_api(
                prompt, image, duration, fps, resolution
            )
    
    def _call_sora_api(
        self,
        prompt: str,
        image: str,
        duration: int,
        fps: int,
        resolution: str
    ) -> Dict[str, Any]:
        """
        Call actual Sora-2 API for video generation.
        
        Note: This is a placeholder for when Sora-2 API becomes available.
        """
        # TODO: Implement actual Sora API call when available
        # For now, fall back to simulation
        logger.info("Sora-2 API not yet available, using simulation")
        return self._simulate_video_generation(
            prompt, image, duration, fps, resolution
        )
    
    def _simulate_video_generation(
        self,
        prompt: str,
        image: str,
        duration: int,
        fps: int,
        resolution: str
    ) -> Dict[str, Any]:
        """
        Simulate video generation for development and testing.
        
        Creates synthetic frames that represent a rotation transformation.
        """
        logger.info("Simulating video generation...")
        
        # Parse resolution
        width, height = map(int, resolution.split('x'))
        
        # Calculate total frames
        total_frames = duration * fps
        
        # Decode base image if provided
        base_image = self._decode_base64_image(image) if image else None
        
        # Generate simulated frames
        frames = []
        for i in range(total_frames):
            # Calculate rotation angle for this frame
            angle = (i / total_frames) * 360
            
            # Create or transform frame
            if base_image:
                frame = self._rotate_image(base_image, angle)
            else:
                frame = self._generate_synthetic_frame(i, total_frames, width, height)
            
            frames.append(frame)
        
        # Simulate processing time
        time.sleep(0.5)  # Simulate API latency
        
        logger.info(f"Generated {len(frames)} frames")
        
        return {
            'frames': frames,
            'metadata': {
                'duration': duration,
                'fps': fps,
                'resolution': resolution,
                'total_frames': total_frames,
                'prompt': prompt,
                'simulation': True
            }
        }
    
    def _decode_base64_image(self, image_data: str) -> np.ndarray:
        """
        Decode base64 image to numpy array.
        """
        # Remove data URL prefix if present
        if image_data.startswith('data:'):
            image_data = image_data.split(',')[1]
        
        # Decode base64
        image_bytes = base64.b64decode(image_data)
        
        # Open with PIL and convert to numpy
        image = Image.open(io.BytesIO(image_bytes))
        return np.array(image)
    
    def _rotate_image(self, image: np.ndarray, angle: float) -> np.ndarray:
        """
        Rotate an image by a given angle.
        """
        # Convert to PIL for rotation
        pil_image = Image.fromarray(image.astype(np.uint8))
        
        # Rotate
        rotated = pil_image.rotate(angle, expand=False, fillcolor=(255, 255, 255))
        
        return np.array(rotated)
    
    def _generate_synthetic_frame(
        self,
        frame_idx: int,
        total_frames: int,
        width: int,
        height: int
    ) -> np.ndarray:
        """
        Generate a synthetic frame for simulation.
        
        Creates a simple geometric shape that rotates.
        """
        # Create blank frame
        frame = np.ones((height, width, 3), dtype=np.uint8) * 255
        
        # Calculate rotation angle
        angle = (frame_idx / total_frames) * 2 * np.pi
        
        # Draw rotating rectangle
        center_x, center_y = width // 2, height // 2
        rect_width, rect_height = width // 3, height // 4
        
        # Calculate rotated corners
        corners = []
        for dx, dy in [(-rect_width//2, -rect_height//2),
                      (rect_width//2, -rect_height//2),
                      (rect_width//2, rect_height//2),
                      (-rect_width//2, rect_height//2)]:
            # Apply rotation
            rx = dx * np.cos(angle) - dy * np.sin(angle)
            ry = dx * np.sin(angle) + dy * np.cos(angle)
            
            # Translate to center
            corners.append((int(center_x + rx), int(center_y + ry)))
        
        # Draw rectangle (simplified - just fill with color)
        # In production, use proper polygon drawing
        for y in range(height):
            for x in range(width):
                # Simple point-in-polygon test
                if self._point_in_rect(x, y, corners):
                    # Color gradient based on rotation
                    r = int(128 + 127 * np.sin(angle))
                    g = int(128 + 127 * np.cos(angle))
                    b = 200
                    frame[y, x] = [r, g, b]
        
        return frame
    
    def _point_in_rect(self, x: int, y: int, corners: list) -> bool:
        """
        Simple point-in-polygon test for rectangle.
        """
        # Simplified test - checks if point is within bounding box
        xs = [c[0] for c in corners]
        ys = [c[1] for c in corners]
        
        return (min(xs) <= x <= max(xs)) and (min(ys) <= y <= max(ys))
    
    def get_video_status(self, video_id: str) -> Dict[str, Any]:
        """
        Get status of video generation job.
        
        Args:
            video_id: Video generation job ID
        
        Returns:
            Status dictionary
        """
        # Placeholder for async video generation
        return {
            'status': 'completed',
            'video_id': video_id,
            'progress': 100
        }
