requirements.txt
openai>=1.6.0
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0
python-dotenv>=1.0.0